Move `Stream` extension methods to own class (getting too long)
