+++ osu.Framework/Extensions/ExtensionMethods.cs
@@ -4,15 +4,12 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Diagnostics;
 using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Security.Cryptography;
 using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
 using osu.Framework.Extensions.ObjectExtensions;
 using osu.Framework.Localisation;
 using osu.Framework.Platform;
@@ -350,93 +347,5 @@ namespace osu.Framework.Extensions
         /// <returns>A <see cref="DisplayMode"/> structure populated with the corresponding properties.</returns>
         internal static DisplayMode ToDisplayMode(this DisplayResolution resolution) =>
             new DisplayMode(null, new Size(resolution.Width, resolution.Height), resolution.BitsPerPixel, (int)Math.Round(resolution.RefreshRate), 0, 0);
-
-        /// <summary>
-        /// Read the full content of a stream.
-        /// </summary>
-        /// <param name="stream">The stream to read.</param>
-        /// <returns>The full byte content.</returns>
-        public static byte[] ReadAllBytesToArray(this Stream stream)
-        {
-            Debug.Assert(stream.Length < int.MaxValue);
-
-            return stream.ReadBytesToArray((int)stream.Length);
-        }
-
-        /// <summary>
-        /// Read the full content of a stream.
-        /// </summary>
-        /// <param name="stream">The stream to read.</param>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>The full byte content.</returns>
-        public static Task<byte[]> ReadAllBytesToArrayAsync(this Stream stream, CancellationToken cancellationToken)
-        {
-            Debug.Assert(stream.Length < int.MaxValue);
-
-            return stream.ReadBytesToArrayAsync((int)stream.Length, cancellationToken);
-        }
-
-        /// <summary>
-        /// Read specified length from current position in stream.
-        /// </summary>
-        /// <param name="stream">The stream to read.</param>
-        /// <param name="length">The length to read.</param>
-        /// <returns>The full byte content.</returns>
-        public static byte[] ReadBytesToArray(this Stream stream, int length)
-        {
-            if (stream is MemoryStream ms && length == ms.GetBuffer().Length)
-                return ms.GetBuffer();
-
-            byte[] buffer = new byte[16 * 1024];
-
-            using (ms = new MemoryStream(length))
-            {
-                int read;
-                int totalRead = 0;
-
-                while ((read = stream.Read(buffer, 0, buffer.Length)) > 0)
-                {
-                    ms.Write(buffer, 0, read);
-                    totalRead += read;
-                }
-
-                if (totalRead != length)
-                    throw new EndOfStreamException();
-
-                return ms.GetBuffer();
-            }
-        }
-
-        /// <summary>
-        /// Read the full content of a stream.
-        /// </summary>
-        /// <param name="stream">The stream to read.</param>
-        /// <param name="length">The length to read.</param>
-        /// <param name="cancellationToken">A cancellation token.</param>
-        /// <returns>The full byte content.</returns>
-        public static async Task<byte[]> ReadBytesToArrayAsync(this Stream stream, int length, CancellationToken cancellationToken)
-        {
-            if (stream is MemoryStream ms && length == ms.GetBuffer().Length)
-                return ms.GetBuffer();
-
-            byte[] buffer = new byte[16 * 1024];
-
-            using (ms = new MemoryStream(length))
-            {
-                int read;
-                int totalRead = 0;
-
-                while ((read = await stream.ReadAsync(buffer.AsMemory(), cancellationToken).ConfigureAwait(false)) > 0)
-                {
-                    await ms.WriteAsync(buffer, 0, read, cancellationToken).ConfigureAwait(false);
-                    totalRead += read;
-                }
-
-                if (totalRead != length)
-                    throw new EndOfStreamException();
-
-                return ms.GetBuffer();
-            }
-        }
     }
 }
